requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Pydantic
pydantic==2.5.0
pydantic-settings==2.1.0

# AWS Services (boto3 only, no aioboto3)
boto3==1.29.7

# AI/ML - Core
groq==0.4.1
scikit-learn==1.3.2
joblib==1.3.2

# Data Processing
pandas==1.5.3
numpy==1.24.3

# Utilities
python-dotenv==1.0.0
colorlog==6.8.0
aiofiles==23.2.1
httpx==0.25.2
cachetools==5.3.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# PostgreSQL/Redshift Driver
psycopg2-binary==2.9.9